Fix a type issue that broke sha1 support on 64-bit arches.
git-svn-id: svn://svn.debian.org/debburn/cdrkit/trunk@805 a95a6be8-091b-0410-adaf-d31e6857962f

+++ genisoimage/sha1.h
@@ -21,7 +21,7 @@
 
 # include <stdio.h>
 
-typedef unsigned long md5_uint32;
+typedef unsigned int md5_uint32;
 
 /* Structure to save state of computation between the single steps.  */
 struct sha1_ctx
